Add files via upload
This project presents a comprehensive analysis of the titles.csv dataset, aiming to explore the ratings of TV series. Using exploratory data analysis (EDA) techniques, the goal is to derive meaningful insights to determine whether the genre and the number of seasons influence the IMDb scores of these series. Additionally, the project seeks to identify the characteristics a new series should have to achieve a rating between 5 and 8 points.